Рефакторинг замечаний
