+++ src/Pages/favorite-page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import OfferList, { OfferData } from '../components/OfferList/offer-list';
-import { Link } from 'react-router-dom';
-import { AppRoutes } from '../constants';
+import Header from '../components/Header/header';
+
 
 type FavoritesPageProps = {
   offers: OfferData[];
@@ -9,39 +9,7 @@ type FavoritesPageProps = {
 
 const FavoritesPage: React.FC<FavoritesPageProps> = ({ offers }) => (
   <div className="page">
-    <header className="header">
-      <div className="container">
-        <div className="header__wrapper">
-          <div className="header__left">
-            <Link className="header__logo-link header__logo-link--active" to={AppRoutes.Root}>
-              <img
-                className="header__logo"
-                src="img/logo.svg"
-                alt="6 cities logo"
-                width="81"
-                height="41"
-              />
-            </Link>
-          </div>
-          <nav className="header__nav">
-            <ul className="header__nav-list">
-              <li className="header__nav-item user">
-                <Link className="header__nav-link header__nav-link--profile" to={AppRoutes.Login}>
-                  <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                  <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                </Link>
-              </li>
-              <li className="header__nav-item">
-                <Link className="header__nav-link" to={AppRoutes.Root}>
-                  <span className="header__signout">Sign out</span>
-                </Link>
-              </li>
-            </ul>
-          </nav>
-        </div>
-      </div>
-    </header>
-
+    <Header/>
     <main className="page__main page__main--favorites">
       <div className="page__favorites-container container">
         <section className="favorites">
