+++ src/Pages/not-found-page.tsx
@@ -1,20 +1,11 @@
 import { AppRoutes } from '../constants';
 import React from 'react';
 import { Link } from 'react-router-dom';
+import Header from '../components/Header/header';
 
 const NotFoundPage: React.FC = () => (
   <div className="page page--gray page--not-found">
-    <header className="header">
-      <div className="container">
-        <div className="header__wrapper">
-          <div className="header__left">
-            <Link className="header__logo-link" to={AppRoutes.Root}>
-              <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-            </Link>
-          </div>
-        </div>
-      </div>
-    </header>
+    <Header />
 
     <main className="page__main page__main--not-found">
       <div className="container">
